# swagger/swagger.yaml
swagger: "2.0"
info:
  version: '1.0.0'
  title: Swagger sam-api
...
  contact:
    email: norbert.bondarczuk@wipro.com
host: localhost:8000
basePath: /api
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /dictionary/segment/{id}:			
    delete:
      description: Removes single segment
      summary: DictionarySegmentDeleteOne
      tags:
      - dictionary-segment
      operationId: DictionarySegmentDeleteOne
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: id
        in: path
        required: true
        type: string
        description: id of segment
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetCount'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'									
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    put:
      description: Changes single segment, all columns
      summary: DictionarySegmentUpdateOne
      tags:
      - dictionary-segment
      operationId: DictionarySegmentUpdateOne
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: id
        in: path
        required: true
        type: string
        description: id of segment
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetCount'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'									
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    patch:
      description: Changes single segment, some columns
      summary: DictionarySegmentUpdateAttributes
      tags:
      - dictionary-segment
      operationId: DictionarySegmentUpdateAttributes
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: id
        in: path
        required: true
        type: string
        description: id of segment
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetCount'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'									
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
definitions:
...
  ResultSetError:
    title: ResultSetError
    type: object
    properties:
      status:
        $ref: '#/definitions/Status12'
      message:
        type: string
...
  ResultSetCount:
    title: ResultSetCount
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
...
  Status:
    title: Status
    example: Ok
    type: string
    enum:
    - Ok
  Status12:
    title: Status12
    example: Error
    type: string
    enum:
    - Error
...